refactor: restructure CursosPage and related components, enhance certificate display and management features
